+++ src/main/java/roomescape/controller/MemberController.java
@@ -1,15 +1,16 @@
-package roomescape.member;
+package roomescape.controller;
 
 import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.net.URI;
+import roomescape.dto.MemberRequest;
+import roomescape.dto.MemberResponse;
+import roomescape.service.MemberService;
 
 @RestController
 public class MemberController {
