+++ src/main/java/roomescape/controller/ViewController.java
@@ -1,10 +1,10 @@
-package roomescape;
+package roomescape.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class PageController {
+public class ViewController {
     @GetMapping("/admin")
     public String admin() {
         return "admin/index";
