+++ src/main/java/roomescape/dao/ThemeDao.java
@@ -1,4 +1,4 @@
-package roomescape.theme;
+package roomescape.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
@@ -6,6 +6,7 @@ import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import roomescape.model.Theme;
 
 @Repository
 public class ThemeDao {
