+++ src/main/java/roomescape/dao/TimeDao.java
@@ -1,4 +1,4 @@
-package roomescape.time;
+package roomescape.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Repository;
 
 import java.sql.PreparedStatement;
 import java.util.List;
+import roomescape.model.Time;
 
 @Repository
 public class TimeDao {
