+++ src/main/java/roomescape/dto/AvailableTime.java
@@ -1,4 +1,4 @@
-package roomescape.time;
+package roomescape.dto;
 
 public class AvailableTime {
     private Long timeId;
