+++ src/main/java/roomescape/dto/ErrorResponse.java
@@ -1,0 +1,3 @@
+package roomescape.dto;
+
+public record ErrorResponse(String message) { }
